build script polishing

+++ packaging/share/os_prereqs/linux_packages/kivy.sh
@@ -9,9 +9,17 @@ SCRIPT_DIR="$( cd -- "$(dirname "$0")" >/dev/null 2>&1 ; pwd -P )"
 cd $SCRIPT_DIR
 source $SCRIPT_DIR/os_package_utils.sh
 
-cd "$(mktemp -d)"
-curl https://raw.githubusercontent.com/kivy/kivy/master/tools/build_linux_dependencies.sh -o build_kivy_deps.sh
+BUILD_DIR=/tmp/EndraBuildKivy
+if ! [ -e $BUILD_DIR ]; then
+    mkdir $BUILD_DIR
+fi
+cd $BUILD_DIR
+
+if ! [ -e build_kivy_deps.sh ]; then
+    curl https://raw.githubusercontent.com/kivy/kivy/master/tools/build_linux_dependencies.sh -o build_kivy_deps.sh
+fi
 chmod +x build_kivy_deps.sh
+
 ./build_kivy_deps.sh
 KIVY_DEPS_ROOT=$(pwd)/kivy-dependencies
 export KIVY_DEPS_ROOT
